Replaced RevenueCatService and SuperwallService with one file
