+++ Examples/UIKit+RevenueCat/Superwall-UIKit+RevenueCat/AppDelegate.swift
@@ -13,7 +13,7 @@ final class AppDelegate: UIResponder, UIApplicationDelegate {
     _ application: UIApplication,
     didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
   ) -> Bool {
-    RevenueCatService.initialize()
+    PaywallManager.configure()
     return true
   }
