+++ Sources/SuperwallKit/Documentation.docc/SuperwallKit.md
@@ -4,7 +4,7 @@ The client SDK for [Superwall](https://superwall.com), a service that lets you r
 
 ## Overview
 
-Superwall makes it easy to show paywalls in your app using UIKit or SwiftUI. If you haven't already, [sign up for a free Superwall account](https://superwall.com/sign-up). Configure the SDK with your API key and present our example paywall in response to a tracked event. When you're ready to use your own paywalls, head to the Superwall dashboard and use the paywall editor to configure them. Then, create a Campaign and define rules for when the paywalls are shown to users.
+Superwall makes it easy to show paywalls in your app using UIKit or SwiftUI. If you haven't already, [sign up for a free Superwall account](https://superwall.com/sign-up). Configure the SDK with your API key and present our example paywall in response to a tracked event. When you're ready to use your own paywalls, head to the Superwall dashboard and use the paywall editor to configure them. Then, create a Campaign and define rules to determine when to show paywalls to users.
 
 ## Topics
